Move versioned clientset into correct package
