+++ pkg/client/clientset_generated/clientset/doc.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// This package is generated by client-gen with arguments: --clientset-path=github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/ --go-header-file=vendor/github.com/kubernetes/repo-infra/verify/boilerplate/boilerplate.go.txt --input=[servicecatalog/v1alpha1] --input-base=github.com/kubernetes-incubator/service-catalog/pkg/apis/
+// This package is generated by client-gen with arguments: --clientset-name=clientset --clientset-path=github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/ --go-header-file=vendor/github.com/kubernetes/repo-infra/verify/boilerplate/boilerplate.go.txt --input=[servicecatalog/v1alpha1] --input-base=github.com/kubernetes-incubator/service-catalog/pkg/apis/
 
-// Package fake has the automatically generated clients.
-package fake
+// This package has the automatically generated clientset.
+package clientset
