+++ pkg/client/clientset_generated/clientset/typed/servicecatalog/v1alpha1/fake/fake_broker.go
