+++ pkg/client/clientset_generated/clientset/typed/servicecatalog/v1alpha1/instance.go
